cleanup: Move FakeClock to its own header
`FakeClock` is a useful construct that only appears in
`service_account_credentials_test.cc`, but I think that it can be used in other
tests as well. I plan on using it in an upcoming PR in a different file.
